requirements.txt
# ===========================================
# VEA CONNECT WEBAPP - REQUIREMENTS.TXT
# ===========================================
# REQUERIDO: Python 3.10+
# Versiones compatibles y validadas

# === DJANGO Y FRAMEWORK PRINCIPAL ===
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
django-widget-tweaks==1.5.0

# === BASE DE DATOS ===
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# === SERVIDOR WEB Y DESPLIEGUE ===
gunicorn==21.2.0
whitenoise==6.6.0

# === VARIABLES DE ENTORNO ===
python-dotenv==1.0.0

# === SERVICIOS DE AZURE ===
azure-communication-messages==1.0.0
azure-communication-phonenumbers==1.0.0
azure-storage-blob==12.19.0
azure-search-documents==11.4.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-ai-formrecognizer==3.3.0
azure-identity==1.15.0
azure-core>=1.35.0,<2.0.0
azure-common==1.1.28

# === CACHE Y REDIS ===
redis==5.0.1
django-redis==5.4.0

# === IA Y PROCESAMIENTO ===
openai==1.3.0

# === UTILIDADES Y PROCESAMIENTO ===
Pillow==10.1.0
requests==2.31.0
numpy==1.24.3
psutil==7.0.0

# === TAREAS ASÍNCRONAS ===
celery==5.3.4
kombu==5.3.4

# === TESTING Y DESARROLLO ===
pytest==7.4.3
pytest-django==4.7.0

# === DEPENDENCIAS ADICIONALES ===
colorlog==6.8.0
python-dateutil==2.8.2
jsonschema==4.20.0
cryptography==41.0.7
urllib3==2.0.7
timeout-decorator==0.5.0
opencensus-ext-azure==1.1.12

# === RESTRICCIONES PARA EVITAR CONFLICTOS ===
setuptools>=65.0.0,<70.0.0
wheel>=0.38.0,<0.42.0
